requirements.txt
anyio==4.6.0
APScheduler==3.10.4
certifi==2024.8.30
et-xmlfile==1.1.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytz==2024.2
six==1.16.0
sniffio==1.3.1
typing-extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
